refactor(flow-editor): streamline imports and enhance node management utilities refactor(edge-panel): remove animation settings for edges refactor(side-panel-flow): adjust z-index for side panel elements
