--- apps/web/hooks/flow/index.ts
@@ -1,1 +1,0 @@
-// Flow hooks have been removed - using experiment-specific flow hooks instead
